fix(hermes): satisfy CI lint — file length + gosec on test DSN
- Split helpers + stubs out of employees_create_test.go into
  employees_helpers_test.go so both files sit under the 300-line cap.
- Trim a blank line in compile_test.go to bring it from 301 → 300.
- Annotate the local-dev DSN constant in compile_test.go with
  //nolint:gosec, matching the convention in
  internal/credentials/integration_test.go.
